+++ requests.cpp
@@ -32,7 +32,7 @@ Json::Dict Bus::Process(const TransportCatalog &db) const {
         dict = {
             {"stop_count",        Json::Node(static_cast<int>(bus->stop_count))},
             {"unique_stop_count", Json::Node(static_cast<int>(bus->unique_stop_count))},
-            {"route_length",      Json::Node(bus->road_route_length)},
+            {"route_length",      Json::Node(static_cast<int>(bus->road_route_length))},
             {"curvature",         Json::Node(bus->road_route_length / bus->geo_route_length)},
         };
     }
